pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mimic-iv-imputation-baselines"
version = "0.1.0"
description = "SSSD & SAITS imputation baselines on MIMIC-IV ICU time series"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "torch>=2.1",
    "pytorch-lightning>=2.1",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "pandas>=2.0",
    "numpy>=1.24",
    "pyarrow>=14.0",
    "scikit-learn>=1.3",
    "wandb>=0.16",
    "tensorboard>=2.15",
    "tqdm>=4.66",
    "einops>=0.7",
    "scipy>=1.11",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.setuptools.packages.find]
include = ["data*", "models*", "metrics*", "baselines_simple*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
